Silence an assert that needlessly crashes at exit
Maybe this assert has a valid purpose, but in practice it seems to only
run at exit, and it fails in that case. I'm not very concerned about the
potential for bugs from not releasing resources during exit.

Removing the assert prevents applications from crash-exiting any time
they try to cleanly exit, which we're seeing especially during system
shutdown.

Signed-off-by: Fernando De La Garza <f.delagarza@trailtech.net>