Overallocate certain VBOs to prevent libGAL crash
In libGAL 6.2.4.p4.0, the computeWLimit function under glDrawElements
overruns certain VBOs when using an index buffer and a draw count
greater than the length of the vertex buffer. The overrun causes a
segfault if that VBO is near a gap in pages mapped by galcore.

Work around this by over-allocating VBOs that will reach the
computeWLimit path to create a safe zone for that overrun. In practice,
this only applies to a few buffers and the memory cost is on the order of
kilobytes.

For testing, set MAPBOX_NO_WLIMIT_WORKAROUND to disable the workaround
and MAPBOX_LOG_WLIMIT_WORKAROUND for additional logging.

Specifically, this is reachable with the dynamicVertexBuffer in a symbol
program, which includes all rendered text. The associated index buffer
is always 1.5x of the vertices, so over-allocating that buffer by 50% is
sufficient. It appears that computeWLimit is only called for certain
VBOs (and the name is a useful hint), so the workaround can be limited
to draw buffers with 3 float components.

The overrun seems to happen near-unconditionally, but it only segfaults
if the buffer is near an unmapped page. galcore maps a large region in
userspace for allocations from the SYSTEM pool, so a segfault is less
common in that case. Buffers in the CONTIGUOUS pool are mapped
individually, so the segfault is likely and becomes more likely over time
with userspace fragmentation. Buffers of around 4000 vertices seem
especially likely to overrun far enough for a crash.

It's unclear if this bug is fixed in newer versions of libGAL, but
upgrading isn't possible at this point because of other bugs and the
current kernel.

%% original patch: 0004-Overallocate-certain-VBOs-to-prevent-libGAL-crash.patch