Process geojson as events on same thread
If the display had a large amount of tracks, routes, and waypoints
the system memory would often be exceed if you imported
a single track, route, or waypoint causing appman-launcher to
be killed by out-of-memory killer (oom killer)

There appears to be an issue with how the Mapbox library manages
memory from geojson sources created on a separate thread. It is
also mentioned in the QMapbox GL documentations "QMapboxGL is
not thread-save and it is assume that it will be accessed from
the same thread as the thread where the OpenGL context lives".

This change gets rid of the threading used to add geojson sources to the map
and instead uses Qt events on the same thread to add the geojson sources.
The importance of the threading was to not block the UI thread and cause
a watchdog timeout. Adding the geojson sources with Qt events still allows
us to achieve this to an extent with the added benefit and better memory
management.

Based on testing using the top command and monitoring
memory usage, this improves memory usage significantly and the system
no longer runs out of memory when importing a single track, route, or waypoint
when there already exists a large amount of tracks, routes, and waypoints.

Testing instructions:

1. Flash the changes onto indianhw-7in
2. Import "GPS Trailmasters Data 2022" from IVI-20259
3. Import a track from "GPX files from Single Chip SNO" from IVI-20259
4. Verify there is no crash and you can pan the map.
5. Pan to a waypoint, track, or route and click on it.
6. Verify you see a popup describing the feature.
7. Change to satellite layer.
8. Verify the maps changes to satellite layer successfully.
9. ssh to the hardare and run the top command.
10. Verify memory usage is below 50% for appman-launcher

Issue: IVI-20259

%% original patch: 0006-Process-geojson-as-events-on-same-thread.patch